Modularized mocha tests in the models dir
